make main.go a playground for now
where I can test simple stuff
