+++ init.go
@@ -1,6 +1,10 @@
 package main
 
 import (
+	"encoding/json"
+	"fmt"
+	"os"
+
 	"github.com/xaaha/hulak/pkg/envparser"
 )
 
@@ -13,4 +17,30 @@ func InitializeProject() {
 	if err != nil {
 		panic(err)
 	}
+}
+
+/*
+Initialize the project and how to substiture a variable.
+This function is just for my dumb brain that forgets how to do simple stuff in a program I wrote
+*/
+func testInitialization() {
+	InitializeProject()
+
+	envMap, err := envparser.GenerateSecretsMap()
+	if err != nil {
+		panic(err)
+	}
+
+	// print entire json
+	niceJson, _ := json.MarshalIndent(envMap, "", "  ")
+	fmt.Println(string(niceJson))
+
+	// how to substitute variable
+	finalAns, err := envparser.SubstitueVariables("env{{PORT}}", envMap)
+	if err != nil {
+		fmt.Println(err)
+	}
+	fmt.Println(finalAns)
+
+	fmt.Println("Default Environment value:", os.Getenv("hulakEnv"))
 }
